+++ app/Filament/Resources/Admin/Dump.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Filament\Resources\Admin;
 
-use App\Filament\Components\TextColumn;
+use App\Filament\Components\Columns\TextColumn;
 use App\Filament\Resources\BaseResource;
 use App\Filament\Resources\Admin\Dump\Pages\CreateDump;
 use App\Filament\Resources\Admin\Dump\Pages\EditDump;
