+++ app/Filament/Resources/Document/Page.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Filament\Resources\Document;
 
-use App\Filament\Components\TextColumn;
+use App\Filament\Components\Columns\TextColumn;
 use App\Filament\Resources\BaseResource;
 use App\Filament\Resources\Document\Page\Pages\CreatePage;
 use App\Filament\Resources\Document\Page\Pages\EditPage;
