+++ app/Filament/Resources/Wiki/Group.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Filament\Resources\Wiki;
 
-use App\Filament\Components\TextColumn;
+use App\Filament\Components\Columns\TextColumn;
 use App\Filament\Resources\BaseResource;
 use App\Filament\Resources\Wiki\Group\Pages\CreateGroup;
 use App\Filament\Resources\Wiki\Group\Pages\EditGroup;
